feat(earnings): extract weekly earnings component
